I'm having a lot of trouble with react so I'm redoing this homework to nail down some basics before moving on to finishing passwords homework and star wars homework
