+++ korilla/src/App.css
@@ -1,29 +1,42 @@
-h1, h2, p {
-  color: black;
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
 }
 
-body {
-  background-color: white;
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
+}
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
 }
 
-.receipt-card {
-  display: flex;
-  flex-direction: row;
-  border: solid 2px orange;
-  margin-left: 20px;
-  width: 300px;
-  padding-left: 10px;
+@keyframes logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
 
-.receipts {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
+@media (prefers-reduced-motion: no-preference) {
+  a:nth-of-type(2) .logo {
+    animation: logo-spin infinite 20s linear;
+  }
 }
 
-h1 {
-  margin-left: 20px;
+.card {
+  padding: 2em;
 }
 
-
-
+.read-the-docs {
+  color: #888;
+}
